Moved CheckRegularExpression() into Utilities.cpp.
